backend/requirement.txt
django=5.0
djangorestframework=3.14.0
Pillow=10.1.0

prakash
123